requirements.txt
anthropic>=0.39.0
Pillow>=10.4.0
requests>=2.32.0
python-dotenv>=1.0.1
schedule>=1.2.2
flask>=3.0.0
